Updated documentation on several files.

+++ ios/Medical Cyborgs/Medical Cyborgs/DeviceConnection.h
@@ -69,7 +69,7 @@
  * or unreadable battery level characteristic.
  */
 
--(NSInteger) batteryLevel;
+-(int) batteryLevel;
 
 
 /**
